pyproject.toml
[project]
name = "qrky"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "qrcode>=8.2",
    "streamlit>=1.46.0",
]
